api/core,bees,cellar,core: Moved NBT related helper classes to api
