+++ src/java/growthcraft/api/core/nbt/INBTSerializable.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License (MIT)
  *
- * Copyright (c) 2015 IceDragon200
+ * Copyright (c) 2015, 2016 IceDragon200
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -21,12 +21,12 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package growthcraft.core.common.nbt;
+package growthcraft.api.core.nbt;
 
 import net.minecraft.nbt.NBTTagCompound;
 
-public interface INbtSerializable
+public interface INBTSerializable
 {
-	void writeToNBT(NBTTagCompound data, String name);
 	void readFromNBT(NBTTagCompound data, String name);
+	void writeToNBT(NBTTagCompound data, String name);
 }
